2026/WebRTC/poc.az.acs2/pyproject.toml
[project]
name = "poc-az-acs2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-communication-identity>=1.5.0",
    "azure-communication-networktraversal>=1.1.0b2",
    "python-dotenv>=1.2.1",
]
